research/pyproject.toml
[project]
name = "quantlib-research"
version = "0.1.0"
description = "Fast quant research on TW stock data — Polars + DuckDB"
requires-python = ">=3.12"
dependencies = [
    "polars>=1.14.0",
    "duckdb>=1.1.0",
    "pyarrow>=17.0.0",
    "typer>=0.12.0",
    "numpy>=2.4.4",
    "alphalens-reloaded>=0.4.6",
    "pyfolio-reloaded>=0.9.9",
    "empyrical-reloaded>=0.5.12",
    "vectorbt>=1.0.0",
    "stockstats>=0.6.8",
    "pyqlib>=0.9.7",
    "xgboost>=3.2.0",
    "catboost>=1.2.10",
    "shioaji>=1.3.3",
]

[tool.uv]
managed = true

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
